+++ package.json
@@ -7,7 +7,7 @@
     "test": "vitest",
     "test:canary": "tsx ./scripts/test-canary.ts",
     "release:cli": "tsx ./scripts/release-cli.ts",
-    "release:cli:upload": "tsx ./scripts/release-cli-upload.ts",
+    "release:cli:upload": "tsx ./scripts/upload-promote.ts",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "prepare": "husky install",
